services/analytics/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
numpy==1.26.0
pandas==2.1.4
scikit-learn==1.3.2
prophet==1.1.5
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.0
plotly==5.18.0
httpx==0.25.2
websockets==12.0
python-dotenv==1.0.0
alembic==1.13.0
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.8